fix(ai): compress Catty requests only after 413
